+++ src/types.d.ts
@@ -16,14 +16,18 @@ export interface ExtensionConfig {
     linkOnOpen: boolean;
     saveFormat: "png" | "svg";
     windowStyle: "macos" | "windows";
+    windowTitleTemplate: string;
 }
 
 export interface ConfigSentToWebview extends ExtensionConfig {
     startLine: number;
-    windowTitle: string;
     tabSize: number;
     fontLigatures: boolean;
     editorID: string;
+    templates: {
+        fileName: string;
+        workspace: string;
+    };
 }
 
 export interface WebviewConfig
@@ -31,6 +35,7 @@ export interface WebviewConfig
     isLocked: boolean;
     isLinked: boolean;
     zoom: number;
+    shouldUpdateTitle: boolean;
 }
 
 export type WebViewConfigKey = keyof WebviewConfig;